+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @swapkit/wallet-keplr
 
+## 1.0.0-rc.8
+
+### Patch Changes
+
+- add scam filter for balances
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.8
+  - @swapkit/helpers@1.0.0-rc.6
+  - @swapkit/types@1.0.0-rc.5
+
 ## 1.0.0-rc.7
 
 ### Patch Changes